mkdocs: add mkdocs-macros-plugin
Add the macros plugin which enables using jinja2 in markdown source
files and provides hooks to inject variables, macros, and filters into
the jinja2 rendering context.

This plugin is injected into the mkdocs plugin list before mdpo which
means that mdpo sees the output of macros when collecting translation
units from the English source. This will allow us to use macros to
generate translation units, but comes with a side effect that some
output may break mdpo. A really easy example of this is that adding
a `{{ macros_info() }}` block produces output which causes mdpo's
initialization code to fall into a recursion loop which eventually
crashes mkdocs.