dev: improve support for Linux dev host
Cross-platform build environments can be a real pain due to OS specific
implementation details. In this Docker Compose based system, we would
like to support at least MacOS and Linux development hosts.

`docker compose` is the Docker Compose version 2.x command name. We are
switching to the `docker-compose` (version 1.x) command as it is more
commonly supported by Linux distribution package managers.

Docker Desktop for MacOS does some nice things to handle effective UID
& GID remapping for mounted volumes. Linux Docker implementations use
volume drivers which do not uniformly provide this support.
A cross-platform trick is to set the runtime user for the Docker
container to match the UID & GID of the host user so that the container
user can write to the mounted volume regardless of volume driver
implementation.