Add helpers for rendering categories
Our content model is based on defining documents in yaml files, tagging
those documents with categories, and providing additional information
about the categories in another collection of yaml files. We use the
mkdocs-macros-plugin to gather up information about the categories and
make that data available for use in markdown pages via a jinja macro.

Usage of the jinja macro as introduced in c8f2dfc was a proof of concept
which showed how to render markdown for a category using jinja. The POC
code as written would require a relatively large amount of boilerplate
per markdown document, and also would make consistent style changes
difficult over time by embedding too much rendering logic in each target
page.

This commit introduces a solution for the boilerplate in the form of
a new local JinjaWrapperPlugin and associated jinja template files. The
wrapper plugin is intended to run before mkdocs-macros-plugin in the
build cycle. The plugin wraps each markdown document in jinja template:

```
{% extends "markdown_base.jinja" %}
{% block markdown %}
... content from .md file goes here ...
{% endblock %}
```

This setup allows us to define imports and other values in our new top
level jinja template so that they are easily used from markdown content.
The initial markdown_base.jinja imports a category.jinja macro template
which encapsulates the boilerplate for rendering a category. With all of
this in place, a markdown file needs only to include a call to the
`category.render` macro to output a single category. All of the logic
for converting a category into markdown source is encapsulated in the
macro which is defined in template rather than python code.

Known issues:
* Local dev only works as expected under Docker Desktop on MacOS with
  runtime UID/GID set to match build time UID/GID of container (same core
  issue as T295823)

Change-Id: Iee772c4319912d3c3b116cce4ca04acc175b8afa