Allow authors to specify an order for items in a category
Document creators want to be able to specify the order in which pieces
of content appear in a page. This change extends processing of the
"category" member of a document YAML description so that it can be
a mapping of category name : sort key values. Legacy lists of category
names are also supported with each element receiving a default sort key
of 0.

Example:
```yaml
title: my cool document
description: this is a cool document
links:
  - url: https://example.org/my_cool_document
    label: project homepage
categories:
    things-that-are-cool: 1
    examples: 11
```

Bug: T299079
Change-Id: I4de7335839fa4ef96ff637d4188c6ab3c5ed10c8