content: Split documentation pages
Splitting out the documentation landing page from the
new contributors guide so that we can link to the
latter when it makes sense. For example, in the tools
section, tool developers would be looking for best
practices for tool docs, not necessarily how to get
started contributing to other docs.

Change-Id: Ie3cbbbe7fd83a24cc815b763200e37d15501edab