plugin: Order language list by display name
Order the list of alternate languages by display name. The order used in
the configuration is also the order seen by the user when selecting
a language. Prior to this patch, the order was simply the dict key
iteration order from the collection of PO files found on disk. That
order is deterministic, but functionally random from the point of view
of a human reading it's output.

Note that we use autonyms for this list, so some sort orders may be
confusing due to visual similarity of various unicode glyphs. A good
example of this is that `Македонски` will sort after `Nederlands`
because the `М` is U+041C (CYRILLIC CAPITAL LETTER EM) and not U+004D
(LATIN CAPITAL LETTER M).

Change-Id: If61d73ac1bc230023f8d8c877cdf885fafe659f3