Remove year from footer copyright statement
Remove the explicit year from the copyright statement used in the site
footer. There is no compelling legal reason to include the year in this
statement as corporate copyright extends 50-70 years from date of
publication for corporate works. Not having a visible year should also
reduce the number of folks from the community who feel compelled to
report the copyright statement as out of date each time the Gregorian
calendar's year changes.

Change-Id: I9d585294162ac06124a63dede89d4285c8b5d2a5