dev: Add test for dirty tree after build
Use `git diff` to check for dirty working tree state following a `mkdocs
build` run. This check is added to both the local 'test' Makefile
target and the 'build' Blubber variant. Pipelinelib config is extended
to run the 'build' Blubber variant as part of the test pipeline.

This change also removes a prior .dockerignore restriction that
prevented most of the `.git` metadata and object database into
containers. We need a working git clone so we can use `git diff`. The
'production' variant container selectively copies content from the
'build' variant in such a way that the `.git` content will not propagate
to the final container.

The general intent of this change is to catch failures to add changes to
data/locale/en/LC_MESSAGES/mkdocs.po resulting from other changes in the
patch during CI.

Bug: T299804
Change-Id: Ia5733cc4310b51d7a96229a36e130801144dadde